correcion de algunos atributos

+++ Dto/Libro/DetalleLibroDigitalRespuestaDto.cs
@@ -4,7 +4,7 @@
     {
         public int IdLibro { get; set; }
         public string? FormatoArchivo { get; set; }
-        public decimal TamanoMB { get; set; }
+        public double TamanoMB { get; set; }
         public string? UrlDescarga { get; set; }
     }
 }
